requirements.txt
ephem==4.0.0.2
python-telegram-bot==13.7
emoji==2.8.0
clarifai-grpc==9.8.0
